[palantir] Tweak models, add mailing capability and report structure
This logic adds the ability for the bot to structure the data it collects and send off to users. We now have nicer modelling of the database, including base functionality for all models (if extending the base class) to interact with the database. These functions also act as integrity guards for data that will be interacting with the database.